requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.50.0
django-debug-toolbar==3.2.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fonttools==4.32.0
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
